+++ packages/libraries/core/CHANGELOG.md
@@ -1,8 +1,24 @@
 # Change Log - @fluentui/react-native
 
-This log was last generated on Mon, 28 Sep 2020 20:32:21 GMT and should not be manually modified.
+This log was last generated on Mon, 19 Oct 2020 18:16:36 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.20.0
+
+Mon, 19 Oct 2020 18:16:36 GMT
+
+### Minor changes
+
+- New experimental component avatar (saadnajmi2@gmail.com)
+
+## 0.19.0
+
+Mon, 12 Oct 2020 04:09:31 GMT
+
+### Minor changes
+
+- shared hooks (email not defined)
 
 ## 0.18.0